Refactor operator debugging implementation
