+++ spark-cypher/src/main/scala/org/opencypher/spark/impl/physical/operators/PhysicalOperatorDebugging.scala
@@ -26,15 +26,13 @@
  */
 package org.opencypher.spark.impl.physical.operators
 
-import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.catalyst.expressions.Attribute
 import org.apache.spark.sql.catalyst.plans.logical.LeafNode
-import org.apache.spark.sql.execution.columnar.InMemoryRelation
 import org.opencypher.spark.api.CAPSSession
 import org.opencypher.spark.configuration.CAPSConfiguration.DebugPhysicalOperators
+import org.opencypher.spark.impl.DataFrameOps._
 import org.opencypher.spark.impl.physical.operators.PhysicalOperatorDebugging.separator
 import org.opencypher.spark.impl.physical.{CAPSPhysicalResult, CAPSRuntimeContext}
-import org.opencypher.spark.impl.util.Profiling.printTiming
 import org.opencypher.spark.impl.{CAPSPatternGraph, CAPSRecords, CAPSUnionGraph}
 
 trait PhysicalOperatorDebugging extends CAPSPhysicalOperator {
@@ -48,7 +46,12 @@ trait PhysicalOperatorDebugging extends CAPSPhysicalOperator {
       println(separator)
       println(s"**$operatorName**")
 
-      val cachedRecordsDf: DataFrame = printExecutionTimingAndPlanThenCacheAndForce("Compute result records", output.records.data)
+      val recordsDf = output.records.data
+      val cachedRecordsDf = {
+        recordsDf.printExecutionTiming("Compute result records")
+        recordsDf.printLogicalPlan
+        recordsDf.cacheAndForce
+      }
       val cachedRecords = CAPSRecords.verifyAndCreate(output.records.header -> cachedRecordsDf)
 
       if (getClass == classOf[ConstructGraph]) {
@@ -57,7 +60,10 @@ trait PhysicalOperatorDebugging extends CAPSPhysicalOperator {
             ug.graphs.collectFirst {
               case (patternGraph: CAPSPatternGraph, _) => patternGraph
             }.map { pg =>
-              printExecutionTimingAndPlanThenCacheAndForce("Compute pattern graph", pg.baseTable.data)
+              val baseTableDf = pg.baseTable.data
+              baseTableDf.printExecutionTiming("Compute pattern graph")
+              baseTableDf.printLogicalPlan
+              baseTableDf.cacheAndForce
             }
         }
       }
@@ -72,35 +78,7 @@ trait PhysicalOperatorDebugging extends CAPSPhysicalOperator {
     }
   }
 
-  /**
-    * Prints computation timing and Spark plan for a DF, then returns a cached and forced version of that DF.
-    */
-  def printExecutionTimingAndPlanThenCacheAndForce(description: String, df: DataFrame): DataFrame = {
-    // Print computation timing
-    printTiming(s"$description") {
-      df.count() // Force evaluation of operator
-    }
-
-    // Print Spark plan
-    println("Spark plan:")
-    val sparkPlan = df.queryExecution.optimizedPlan
-    implicit val sparkSession = df.sparkSession
-    // Remove cached inputs from plan
-    val planWithoutCached = sparkPlan.transformDown {
-      case _: InMemoryRelation => CachedOperatorInput
-      case other => other
-    }
-    println(planWithoutCached.treeString)
-
-    // Cache
-    val cachedDf = df.cache()
-
-    // Force
-    cachedDf.count() // Force evaluation of cached DF
-    cachedDf
   }
-
-}
 
 object PhysicalOperatorDebugging {
